YUK-9484 Porter Response Monitor - server side messaging and support Improved startup behavior, added exception handling around most CMS and AMQ functions

+++ yukon-server/Message/include/PorterResponseMessage.h
@@ -14,7 +14,7 @@ class IM_EX_MSG PorterResponseMessage : public StreamableMessage
     long _connectionId;
     long _deviceId;
     long _userMessageId;
-    long _status;
+    int  _status;
     bool _final;
 
 public:
